+++ .github/workflows/CHANGELOG.md
@@ -2,4 +2,11 @@
 
 ## What's changed
 
-**issue #76 include `@props` directive**
+** issue #77 spatie/laravel-honeypot directive not loading**
+
+- include spatie `@honeypot` directive by default
+- fix config scanning inside callbacks
+
+** components resource file**
+
+- show declaration finder for component resource file